Gjort innsending av skjemaer mer robust ved å legge til disabled flagg. Skrevet om skjemasteg til styled component.
