+++ src/frontend/index.less
@@ -27,7 +27,6 @@
 
 
 // Felles
-@import "./komponenter/Felleskomponenter/Skjemasteg/skjemasteg.less";
 @import "./komponenter/Felleskomponenter/Informasjonsbolk/informasjonsbolk.less";
 @import "./komponenter/Felleskomponenter/InputMedLabelTilVenstre/inputmedlabeltilvenstre.less";
 @import "./komponenter/Felleskomponenter/Hendelsesoversikt/hendelsesoversikt.less";
