--- src/frontend/komponenter/Felleskomponenter/Skjemasteg/skjemasteg.less
@@ -1,17 +1,0 @@
-.skjemasteg {
-    padding: 2rem;
-
-    .typo-innholdstittel {
-        padding-bottom: 1rem;
-    }
-
-    &__navigering {
-        padding: 1rem 0;
-        display: flex;
-        flex-direction: row-reverse;
-
-        &--flex {
-            flex: 1;
-        }
-    }
-}
